Created separate read-file and write-file modules and removed unnecessary try catch statements.
