+++ node-notes-cli/app.js
@@ -4,10 +4,16 @@ import deleteNote from './delete.js';
 import updateNote from './update.js';
 
 const [,, feature, ...args] = process.argv;
-switch (feature) {
-  case 'read': readNotes(1); break;
-  case 'add': addNote(args); break;
-  case 'delete': deleteNote(args); break;
-  case 'update': updateNote(args[0], args.slice(1)); break;
-  default: console.error(`${feature} is not a feature!`); process.exit(1);
+
+try {
+  switch (feature) {
+    case 'read': await readNotes(); break;
+    case 'add': await addNote(args); break;
+    case 'delete': await deleteNote(args); break;
+    case 'update': await updateNote(args[0], args.slice(1)); break;
+    default: console.error(`${feature} is not a feature!`); process.exit(1);
+  }
+} catch (err) {
+  console.error(err.message);
+  process.exit(1);
 }
