+++ node-notes-cli/data.json
@@ -1,5 +1,5 @@
 {
-  "nextId": 17,
+  "nextId": 18,
   "notes": {
     "11": "The event loop is how a JavaScript runtime pushes asynchronous callbacks onto the stack once the stack is cleared.",
     "12": "Prototypal inheritance is how JavaScript objects delegate behavior.",
