+++ node-notes-cli/delete.js
@@ -1,20 +1,15 @@
-import * as fs from 'node:fs/promises';
-import readNotes from './read.js';
+import read from './read-file.js';
+import write from './write-file.js';
 
 export default async function deleteNote(ids) {
-  try {
-    const dataCopyObject = await readNotes(0);
-    ids.forEach((id) => {
-      if (dataCopyObject.notes[id]) {
-        delete dataCopyObject.notes[id];
-      } else {
-        console.error(`Note ${id} doesn't exist!`);
-        process.exit(1);
-      }
-    });
-    await fs.writeFile('data.json', JSON.stringify(dataCopyObject, null, 2));
-  } catch (err) {
-    console.error(err.message);
-    process.exit(1);
-  }
+  const dataCopyObject = await read();
+  ids.forEach((id) => {
+    if (dataCopyObject.notes[id]) {
+      delete dataCopyObject.notes[id];
+    } else {
+      console.error(`Note ${id} doesn't exist!`);
+      process.exit(1);
+    }
+  });
+  await write(dataCopyObject);
 }
